requirements.txt
Django==2.2.4
mysqlclient=1.4.4
pip==19.0.3
pytz==2019.2
setuptools==40.8.0
sqlparse==0.3.0
Pillow==6.2.0